refactor(switch): ♻️  move switch parts into separate files
